#528 coordinator configurator cleanup
